requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
DateTime==5.1
deprecation==2.1.0
easyocr==1.7.0
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.3.3
future==0.18.3
gensim==4.3.1
gevent==23.9.0.post1
greenlet==2.0.2
grpcio==1.56.2
idna==3.4
imageio==2.31.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.40.1
magic-filter==1.0.9
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
ninja==1.11.1
numba==0.57.1
numpy==1.24.4
openai==0.27.8
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==4.23.4
pyaes==1.6.1
pyasn1==0.5.0
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.11
pydub==0.25.1
pyparsing==3.0.9
PySoundFile==0.9.0.post1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
soupsieve==2.4.1
SpeechRecognition==3.10.0
sympy==1.12
Telethon==1.28.5
tifffile==2023.8.25
tinkoff==0.1.1
tinkoff-investments==0.2.0b58
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
websocket==0.2.1
websockets==11.0.3
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
